requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
africastalking==1.2.9
# python-dateutil==2.8.2
geopy==2.3.0
pytz==2023.3
# boto3==1.28.58
pytest==7.4.3
pytest-cov==4.1.0
